+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Gitaly changelog
 
+UNRELEASED
+
+- Implement recursive tree entries fetching
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/600
+
 [v0.84.0](https://gitlab.com/gitlab-org/gitaly/tags/v0.84.0)
 
 - Send gitaly-ruby exceptions to Sentry